fix create projects names
